% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colEquivalence.R
\name{colEquivalence}
\alias{colEquivalence}
\title{colEquivalence}
\usage{
colEquivalence(x)
}
\arguments{
\item{x}{a 2D matrix of column-wise component labels that is returned by colPass}
}
\value{
list containing voxel indices (arr.ind=FALSE type) of connected components resulting from the first pass (column-wise) of the algorithm on a single slice
}
\description{
internal function used by ccl
}
\examples{
\dontrun{

}
}

