% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colPass.R
\name{colPass}
\alias{colPass}
\title{colPass}
\usage{
colPass(x)
}
\arguments{
\item{x}{a 2D binary matrix, e.g., single slice from lesion mask}
}
\value{
matrix of connected components labeled 1, 2, ..., where each component has a unique label
}
\description{
internal function used by ccl
}
\examples{
\dontrun{

}
}

