% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroups.R
\name{getGroups}
\alias{getGroups}
\title{getGroups}
\usage{
getGroups(li, ve, l, os)
}
\arguments{
\item{li}{original slice-wise label image}

\item{ve}{voxel equivalence classes}

\item{l}{length of ve}

\item{os}{number of pixels in one slice}
}
\value{
list of slice-wise groups and all lower connected slice groups
}
\description{
internal function used by ccl
}
\examples{
\dontrun{

}
}

