% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowPass.R
\name{rowPass}
\alias{rowPass}
\title{rowPass}
\usage{
rowPass(x, eq)
}
\arguments{
\item{x}{the 2D binary matrix output by colPass}

\item{eq}{the list of voxel equivalences output by colEquivalence}
}
\value{
2D binary matrix of connected component labels
}
\description{
internal function used by ccl
}
\examples{
\dontrun{

}
}

