% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSlice.R
\name{singleSlice}
\alias{singleSlice}
\title{singleSlice}
\usage{
singleSlice(x)
}
\arguments{
\item{x}{a 2D binary matrix, e.g., single slice of lesion mask}
}
\value{
matrix with labeled components
}
\description{
internal function used by ccl
}
\examples{
\dontrun{

}
}

