% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicePass.R
\name{slicePass}
\alias{slicePass}
\title{slicePass}
\usage{
slicePass(x, verbose = TRUE)
}
\arguments{
\item{x}{a 3D binary matrix, e.g., lesion mask}

\item{verbose}{if TRUE prints progress bars}
}
\value{
equi a list of voxel equivalencies defined by the labels
}
\description{
internal function used by ccl
}
\examples{
\dontrun{

}
}

