% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecPass.R
\name{vecPass}
\alias{vecPass}
\title{vecPass}
\usage{
vecPass(x)
}
\arguments{
\item{x}{a binary vector, i.e., a single column from a single lesion mask slice}
}
\value{
vector of connected components labeled 1, 2, ...
}
\description{
internal function used by ccl
}
\examples{
\dontrun{

}
}

